Got the gui picker working "RBY Interactive", also it now removes the rby warning as well. Adds a note explaining about the yetis.
The images seem to work in 1.10.4, going to try to test in 1.11 now.

Opportunities to improve further -- make it read the custom file?
Make it read the map pool from the cfg files rather than hard coded lists?

Allow 2v2 interactive selection?

Find a way to make it work with default era random?